imp(EMP-2140): Improve component creation, deletion and display (#15)

+++ index.js
@@ -4,17 +4,40 @@ import { Command } from "commander";
 import chalk from "chalk";
 import fs from "fs-extra";
 import path from "path";
+import { input, select, number, expand } from "@inquirer/prompts";
 
-const program = new Command();
 const configPath = path.join(process.env.HOME, ".ecli", "config.json");
 const componentsDir = path.join(process.cwd(), "Components");
 
-async function fetchComponent(config, componentName) {
-  const url = `${config.apiUrl}/workflows/name/${componentName}`;
+async function fetchGetComponent(config, name) {
+  const url = `${config.apiUrl}/workflows/name/${name}`;
   return fetch(url).then((res) => res.json());
 }
 
-async function updateComponent(config, id, data) {
+async function fetchGetServers(config) {
+  const url = `${config.apiUrl}/servers`;
+  return fetch(url).then((res) => res.json());
+}
+
+async function fetchCreateComponent(config, data) {
+  const url = `${config.apiUrl}/workflows`;
+  return fetch(url, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(data),
+  }).then((res) => res.json());
+}
+
+async function fetchRemoveComponent(config, id) {
+  const url = `${config.apiUrl}/workflows/${id}`;
+  return fetch(url, {
+    method: "DELETE",
+  }).then((res) => res.json());
+}
+
+async function fetchUpdateComponent(config, id, data) {
   const url = `${config.apiUrl}/workflows/${id}`;
   return fetch(url, {
     method: "PUT",
@@ -34,18 +57,79 @@ async function initConfig() {
   console.log(chalk.green("Configuration file created successfully!"));
 }
 
-async function addComponent(componentName) {
-  const componentPath = path.join(componentsDir, componentName);
+async function newComponent() {
+  const config = await fs.readJson(configPath);
 
+  const { data: servers, error: fetchServersError } =
+    await fetchGetServers(config);
+  if (fetchServersError != null) {
+    console.error(chalk.red(`Failed to fetch servers: ${fetchServersError}`));
+    return;
+  }
+  const serverChoices = servers.map((server) => ({
+    name: server.name,
+    value: server.id,
+  }));
+
+  const componentData = {
+    name: await input({
+      message: "Enter the name of the component",
+      required: true,
+    }),
+    label: await input({
+      message: "Enter the label of the component",
+      required: true,
+    }),
+    description: await input({
+      message: "Enter the description of the component",
+      required: true,
+    }),
+    server_id: await select({
+      message: "Select the server",
+      choices: serverChoices,
+    }),
+    output_mime_type: await input({
+      message: "Enter the output mime type",
+      initial: "image/png",
+    }),
+    type: await select({
+      message: "Select the type of the component",
+      choices: [
+        { title: "Basic", value: "basic" },
+        { title: "Comfy Workflow", value: "comfy_workflow" },
+      ],
+    }),
+    order: await number({
+      message: "Enter the order of the component",
+      default: 0,
+    }),
+    display: false,
+  };
+
+  // Check if the component already exists
+  const componentPath = path.join(componentsDir, componentData.name);
   if (await fs.pathExists(componentPath)) {
-    console.error(chalk.red(`Component "${componentName}" already exists!`));
+    console.error(
+      chalk.red(`Component "${componentData.name}" already exists!`)
+    );
+    return;
+  }
+
+  const { error: createComponentError } = await fetchCreateComponent(
+    config,
+    componentData
+  );
+  if (createComponentError != null) {
+    console.createComponentError(
+      chalk.red(`Failed to create component: ${createComponentError}`)
+    );
     return;
   }
 
   await fs.ensureDir(componentPath);
   await fs.mkdir(path.join(componentPath, "ref"));
-  await fs.writeFile(path.join(componentPath, "api.json"), "");
-  await fs.writeFile(path.join(componentPath, "form.json"), "");
+  await fs.writeFile(path.join(componentPath, "api.json"), "{}");
+  await fs.writeFile(path.join(componentPath, "form.json"), "{}");
   await fs.writeFile(
     path.join(componentPath, "credits.js"),
     `function computeCredits(context) {
@@ -53,17 +137,36 @@ async function addComponent(componentName) {
 }`
   );
 
-  console.log(chalk.green(`Component "${componentName}" added successfully!`));
+  console.log(
+    chalk.green(`Component "${componentData.name}" added successfully!`)
+  );
 }
 
 async function removeComponent(componentName) {
+  const config = await fs.readJson(configPath);
+  const { data: component, error: getComponentError } = await fetchGetComponent(
+    config,
+    componentName
+  );
+  if (getComponentError != null) {
+    console.error(chalk.red(`Failed to fetch component: ${getComponentError}`));
+    return;
+  }
+  const { error: removeComponentError } = await fetchRemoveComponent(
+    config,
+    component.id
+  );
+  if (removeComponentError != null) {
+    console.error(
+      chalk.red(`Failed to remove component: ${removeComponentError}`)
+    );
+    return;
+  }
   const componentPath = path.join(componentsDir, componentName);
-
   if (!(await fs.pathExists(componentPath))) {
     console.error(chalk.red(`Component "${componentName}" does not exist!`));
     return;
   }
-
   await fs.remove(componentPath);
   console.log(
     chalk.green(`Component "${componentName}" removed successfully!`)
@@ -73,7 +176,7 @@ async function removeComponent(componentName) {
 async function applyComponents(componentName) {
   const config = await fs.readJson(configPath);
 
-  const { data: component, error: componentError } = await fetchComponent(
+  const { data: component, error: componentError } = await fetchGetComponent(
     config,
     componentName
   );
@@ -123,9 +226,13 @@ async function applyComponents(componentName) {
     workflow: api.workflow,
     credits_script: credits,
   };
-  const { error: updateError } = await updateComponent(config, component.id, {
-    data,
-  });
+  const { error: updateError } = await fetchUpdateComponent(
+    config,
+    component.id,
+    {
+      data,
+    }
+  );
 
   if (updateError != null) {
     console.error(chalk.red(`Failed to update component "${componentName}"!`));
@@ -140,7 +247,7 @@ async function applyComponents(componentName) {
 async function getComponent(componentName, options) {
   console.log(`Getting details of component "${componentName}"...`);
   const config = await fs.readJson(configPath);
-  const { data: component, error } = await fetchComponent(
+  const { data: component, error } = await fetchGetComponent(
     config,
     componentName
   );
@@ -173,6 +280,69 @@ async function getComponent(componentName, options) {
   }
 }
 
+async function displayComponents(componentName) {
+  const config = await fs.readJson(configPath);
+  const { data: components, error: getComponentError } =
+    await fetchGetComponent(config, componentName);
+  if (getComponentError != null) {
+    console.error(
+      chalk.red(`Failed to fetch components: ${getComponentError}`)
+    );
+    return;
+  }
+
+  const answer = await expand({
+    message: `Do you want to display/hide the ${componentName}?`,
+    default: "n",
+    choices: [
+      {
+        key: "y",
+        name: "Display",
+        value: "yes",
+      },
+      {
+        key: "n",
+        name: "Hide",
+        value: "no",
+      },
+      {
+        key: "x",
+        name: "Abort",
+        value: "abort",
+      },
+    ],
+  });
+
+  switch (answer) {
+    case "yes": {
+      const { error } = await fetchUpdateComponent(config, components.id, {
+        display: true,
+      });
+      if (error != null) {
+        console.error(chalk.red(`Failed to update component: ${error}`));
+        return;
+      }
+      console.log(chalk.green("Component displayed successfully!"));
+      break;
+    }
+    case "no": {
+      const { error } = await fetchUpdateComponent(config, components.id, {
+        display: false,
+      });
+      if (error != null) {
+        console.error(chalk.red(`Failed to update component: ${error}`));
+        return;
+      }
+      console.log(chalk.green("Component hidden successfully!"));
+      break;
+    }
+    case "abort":
+      break;
+  }
+}
+
+const program = new Command();
+
 program
   .command("init")
   .description("Initialize the configuration file")
@@ -184,10 +354,9 @@ const componentsCommand = program
   .description("Manage components");
 
 componentsCommand
-  .command("add")
+  .command("new")
   .description("Add a new component")
-  .arguments("<name>")
-  .action(addComponent);
+  .action(newComponent);
 
 componentsCommand
   .command("remove")
@@ -200,6 +369,12 @@ componentsCommand
   .description("Apply the components")
   .arguments("<name>")
   .action(applyComponents);
+
+componentsCommand
+  .command("display")
+  .description("Display the components")
+  .arguments("<name>")
+  .action(displayComponents);
 
 componentsCommand
   .command("get")
